リファクタ作業2
