+++ frontend/src/features/auth/graphql/mutations/registerUser.graphql
