+++ frontend/src/features/auth/graphql/queries/authQueries.graphql
