+++ frontend/src/features/auth/graphql/queries/getMe.graphql
