+++ frontend/src/features/todo/graphql/mutations/createCategory.graphql
