+++ frontend/src/features/todo/graphql/mutations/deleteCategory.graphql
