+++ frontend/src/features/todo/graphql/mutations/deleteTodo.graphql
