+++ frontend/src/features/todo/graphql/queries/getCategories.graphql
